+++ src/routes/Quizzes/Quiz.css
@@ -6,7 +6,7 @@
 .parallel {
 	display: flex;
 	flex-direction: row;
-	justify-content: space-evenly;
+	justify-content: center;
 	align-items: baseline;
 }
