+++ src/routes/Quizzes/Quiz1.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import {
   Switch,
-  Route,
-  NavLink
+  Route
 } from "react-router-dom";
 import Quiz2 from './Quiz2';
-import PuzzleImg from '../../components/PuzzleImg';
+import Questions from '../../components/Questions';
 import './Quiz.css';
 
 
@@ -96,126 +95,13 @@ const questions = [
 ]
 
 class Quiz1 extends React.Component {
-  state = {
-    submitted: false,
-    correct: false,
-    giveUp: false,
-    q: 0,
-    quizComplete: false
-  };
-
-  simplifyString(theirAnswer) {
-    let simpleAnswer = theirAnswer.toLowerCase()
-    simpleAnswer = simpleAnswer.replace(/ /g, '');
-    simpleAnswer = simpleAnswer.replace(/-/g, '');
-    simpleAnswer = simpleAnswer.replace(/\./g, '');
-    simpleAnswer = simpleAnswer.replace(/!/g, '');
-
-    return simpleAnswer;
-  };
-
-  checkAnswer() {
-    const { q } = this.state;
-    const theirAnswer = document.getElementById('answer').value;
-    if (questions[q].possibilities.includes(this.simplifyString(theirAnswer))) {
-      this.setState({correct: true});
-    }
-    this.setState({submitted: true});
-    questions[q].attempts = questions[q].attempts + 1;
-  }
-
-  nextQuestion() {
-    const { q } = this.state;
-    if (q < 9) {
-      this.setState({
-        submitted: false,
-        correct: false,
-        giveUp: false,
-        q: q + 1
-      });
-    } else {
-      this.setState({
-        quizComplete: true
-      })
-    }
-  }
-
-  tryAgain() {
-    this.setState({submitted: false})
-  }
-
-  showAnswer() {
-    this.setState({giveUp: true})
-  }
-
-  ordinalSuffix(i) {
-    var j = i % 10;
-    var k = i % 100;
-    if (i === 0) {
-        return "";
-    }
-    if (j === 1 && k !== 11) {
-        return i + "st";
-    }
-    if (j === 2 && k !== 12) {
-        return i + "nd";
-    }
-    if (j === 3 && k !== 13) {
-        return i + "rd";
-    }
-    return i + "th";
-};
+  
 
   render() {
-    const { submitted, correct, giveUp, q, quizComplete } = this.state;
     return (
       <div className="contentContainer">
         <div className="mainContent" id="questions">
-          {
-            (!quizComplete) &&
-            <div>
-              <h1 id="questionNumber">Quiz 1 - Question {q + 1}</h1>
-              <PuzzleImg id="questionImg" src={questions[q].src} alt={questions[q].alt}></PuzzleImg>
-            </div>
-          }
-          {
-            (!submitted && !giveUp && !quizComplete) &&
-            <div class="parallel">
-              <input type="text" id="answer" name="answer" placeholder="Your answer"></input>
-              <button className="styledButton" onClick={() => this.checkAnswer()}>Submit</button>
-            </div>
-          }
-          {
-            (submitted && correct && !giveUp && !quizComplete) &&
-            <div class="parallel">
-              <h4>Correct! The answer is "{questions[q].correct}".</h4>
-              <button className="styledButton" onClick={() => this.nextQuestion()}>Next</button>
-            </div>
-          }
-          {
-            (submitted && !correct && !giveUp && !quizComplete) &&
-            <div class="parallel">
-              <h4>Incorrect... that was your {this.ordinalSuffix(questions[q].attempts)} attempt.</h4>
-              <button className="styledButton" onClick={() => this.tryAgain()}>Try again</button>
-              <button className="styledButton" onClick={() => this.showAnswer()}>Show answer</button>
-            </div>
-          }
-          {
-            (submitted && giveUp && !quizComplete) &&
-            <div class="parallel">
-              <h4>The answer is "{questions[q].correct}".</h4>
-              <button className="styledButton" onClick={() => this.nextQuestion()}>Next</button>
-            </div>
-          }
-          {
-            (quizComplete) &&
-            <div>
-              <h4>Congratulations! You have completed Quiz 1.</h4>
-              <NavLink activeClassName="activeLink" exact to="/quiz2">
-                <button className="styledButton" id="tryQuiz">Try Quiz 2</button>
-              </NavLink>
-            </div>
-          }
+          <Questions questions={questions} quizNumber={1} nextQuizNumber={2}></Questions>
 
 
           <Switch>
